Make arrays_intersect a static protected method in PHP_Typography
