+++ tests/class-php-typography-functions-test.php
@@ -22,6 +22,8 @@
  *  @license http://www.gnu.org/licenses/gpl-2.0.html
  */
 
+namespace PHP_Typography\Tests;
+
 // Can't autoload functions.
 require_once dirname( __DIR__ ) . '/php-typography/php-typography-functions.php';
 
@@ -44,53 +46,6 @@ class PHP_Typography_Functions_Test extends \PHPUnit\Framework\TestCase {
 	}
 
 	/**
-	 * Provide data for testing arrays_intersect.
-	 *
-	 * @return array
-	 */
-	public function provide_arrays_intersect_data() {
-		return [
-			[ [], [], false ],
-			[ [ 1, 2, 3 ], [ 2, 4, 1 ], true ],
-			[ [ 1, 2, 3 ], [], false ],
-			[ [], [ 1, 2, 3 ], false ],
-		];
-	}
-
-	/**
-	 * $a1 and $a2 need to be arrays of object indexes < 10
-	 *
-	 * @covers \PHP_Typography\arrays_intersect
-	 * @dataProvider provide_arrays_intersect_data
-	 *
-	 * @param  array $a1     First array.
-	 * @param  array $a2     Second array.
-	 * @param  bool  $result Expected result.
-	 */
-	public function test_arrays_intersect( array $a1, array $a2, $result ) {
-		$nodes = [];
-		for ( $i = 0; $i < 10; ++$i ) {
-			$nodes[] = new \DOMText( "foo $i" );
-		}
-
-		$array1 = [];
-		foreach ( $a1 as $index ) {
-			if ( isset( $nodes[ $index ] ) ) {
-				$array1[] = $nodes[ $index ];
-			}
-		}
-
-		$array2 = [];
-		foreach ( $a2 as $index ) {
-			if ( isset( $nodes[ $index ] ) ) {
-				$array2[ spl_object_hash( $nodes[ $index ] ) ] = $nodes[ $index ];
-			}
-		}
-
-		$this->assertSame( $result, \PHP_Typography\arrays_intersect( $array1, $array2 ) );
-	}
-
-	/**
 	 * Test get_object_hash function.
 	 *
 	 * @covers \PHP_Typography\get_object_hash
@@ -100,7 +55,7 @@ class PHP_Typography_Functions_Test extends \PHPUnit\Framework\TestCase {
 		$this->assertInternalType( 'string', $hash1 );
 		$this->assertGreaterThan( 0, strlen( $hash1 ) );
 
-		$hash2 = \PHP_Typography\get_object_hash( new stdClass() );
+		$hash2 = \PHP_Typography\get_object_hash( new \stdClass() );
 		$this->assertInternalType( 'string', $hash2 );
 		$this->assertGreaterThan( 0, strlen( $hash2 ) );
